rrr: convert reactor/epoll_wrapper to module (rrr.epoll_wrapper)
epoll_wrapper.cc becomes the rrr.epoll_wrapper module interface unit;
epoll_wrapper.h deleted. Contents: `Pollable` interface, `Epoll` class
(with templated `Wait<ReadyHandler>`), `PollMode`/`PollReady`
namespaces. The non-templated `Epoll::Wait()` overload (formerly
out-of-line in .cc) lives in the non-export `namespace rrr { }` tail
of the module.

Includers (5) updated to `import rrr.epoll_wrapper;`:
  - src/rrr/reactor/reactor.h, src/rrr/rpc/{client,server}.hpp,
    src/rrr/rpc/pollable_proxy.h, src/rrr/rrr.hpp.

Dropped: the unused `class PollThreadWorker;` forward-decl inside the
module's `export namespace rrr { }` block. reactor.h still
forward-declares `PollThreadWorker` in the global module, and having
the same name forward-declared in two attachments triggered "follows
declaration in module rrr.epoll_wrapper" errors across every TU that
included reactor.h. The decl wasn't referenced by anything in this
module anyway.

Measurement (rrr + rpcbench, -j32, clean):
  +rand:           38.35 s, 161.0 MB PCM
  +epoll_wrapper:  39.01 s, 167.1 MB PCM, librrr.a 9.66 MB

+0.7 s wallclock; 2.31× baseline. 3× threshold (~50 s) still has
margin. PCM is now 10 files / 167 MB — the gradient is plateauing.
