+++ docs/dev/srpc_module_migration_plan.md
@@ -145,4 +145,5 @@ build && cmake -G Ninja -B build ... && ninja rrr rpcbench`).
 | base/misc (no-shim) | 30.14 | 7 | 112.4 | 10.34 | misc.hpp deleted; 3 includers updated. debugging+logging migrated from GMF forward-decls to `import rrr.misc;` (now that misc is a module). BMI dedup visible: rrr.logging.pcm is only 0.6 MB because it just references other module BMIs; rrr.debugging.pcm dropped from 25.7 → 24.6 MB. Dropped unused macros: `arraysize`, `TIME_NOW_STR_SIZE`, `ArraySizeHelper` template. Wallclock +40% jump in this step — more TUs now consume more module BMIs transitively. |
 | base/threading (no-shim) | 37.01 | 8 | 136.8 | 9.99 | threading.hpp deleted; 9 includers updated. ~1000-line merged file (largest so far): Pthread wrappers, SpinLock/SpinMutex family, SpinCondVar, Queue<T>, ThreadPool, RunLater. Switched `max(...)` → `std::max(...)` (lost via `using namespace std;` drop). 2.19× baseline; threshold (3×) remains. |
 | misc/rand (no-shim) | 38.35 | 9 | 161.0 | 10.15 | rand.hpp deleted; only rrr.hpp included it. RandomGenerator class (all static methods). +1.3 s wallclock; 2.27× baseline. |
+| reactor/epoll_wrapper (no-shim) | 39.01 | 10 | 167.1 | 9.66 | epoll_wrapper.h deleted; 5 includers updated. `Pollable` interface + `Epoll` class with templated `Wait<>` member. Dropped the unused `class PollThreadWorker;` forward-decl (was causing global-vs-module attachment clash in reactor.h's similar forward-decl). |
