+++ src/rrr/CMakeLists.txt
@@ -24,6 +24,7 @@ set(RRR_MODULE_SRC
     ${CMAKE_CURRENT_SOURCE_DIR}/base/strop.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/base/threading.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/misc/rand.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/reactor/epoll_wrapper.cc
 )
 list(REMOVE_ITEM RRR_BENCH_SRC ${RRR_MODULE_SRC})
