+++ src/rrr/rpc/client.hpp
@@ -25,8 +25,9 @@
 
 #include "../base/all.hpp"
 #include "../misc/marshal.hpp"
-#include "../reactor/epoll_wrapper.h"
 #include "../reactor/reactor.h"
+
+import rrr.epoll_wrapper;
 
 
 #include "channel.hpp"
